pyproject.toml
[project]
name = "chm_utils"
version = "0.0.1"
dependencies = [
  "boto3",
  "PyYAML"
]
requires-python = ">=3.8"
authors = [
  {name = "Adrian Oesch", email = "helloadrianoesch@chmedia.ch"},
]
description = "A custom python package with common utilities!"
readme = "readme.md"

[project.optional-dependencies]
snowflake = [
  "snowflake-connector-python>=3.0.2",
  "numpy<2",
  "pandas>=2.0.1"
]
redis = [
  "redis"
]

[project.scripts]
chm_utils = "chm_utils.cli:main"